feat: WebSocket 文件流式传输 + 虚拟滚动渲染 (v2.3.0)
文档浏览器从 HTTP GET 全量加载改为 WebSocket 0x05 二进制帧流式传输，
突破 10MB 文件大小限制，支持 50MB 以内所有类型文件。

传输层:
- 新增 WS 二进制帧 0x05 (file-chunk) + JSON 控制消息 (stream-file/end/error)
- 服务端 createReadStream 64KB 分块，背压控制防止 WS 缓冲区溢出
- 客户端 fileStreamBus 事件总线桥接 TerminalView(WS) ↔ PlanPanel(渲染)

渲染层:
- text/log: react-window FixedSizeList 虚拟滚动，边传边显示
- md/html: 流式累积完成后交给现有渲染器
- pdf: 二进制流直传 Uint8Array，消除 base64 的 33% 体积膨胀
- 状态栏进度条 200ms 刷新 (5fps)

+++ server/src/auth.ts
@@ -6,5 +6,5 @@ export function safeTokenCompare(a: string, b: string): boolean {
   const key = 'ai-cli-online-token-compare';
   const hmacA = createHmac('sha256', key).update(a).digest();
   const hmacB = createHmac('sha256', key).update(b).digest();
-  return timingSafeEqual(hmacA, hmacB);
+  return timingSafeEqual(hmacA as Uint8Array, hmacB as Uint8Array);
 }
